refactor(journeys-admin): rename CheckBoxOption dir to CheckboxOption (ENG-3725 A3)
Directory casing now matches the CheckboxOption component and files it
contains.
