+++ apps/journeys-admin/src/components/JourneyVisitorsList/FilterDrawer/ExportDialog/FilterForm/FilterForm.tsx
@@ -14,7 +14,7 @@ import ChevronUp from '@core/shared/ui/icons/ChevronUp'
 
 import { EventType } from '../../../../../../__generated__/globalTypes'
 import { FILTERED_EVENTS } from '../../../../../libs/useJourneyEventsExport/utils/constants'
-import { CheckboxOption } from '../CheckBoxOption'
+import { CheckboxOption } from '../CheckboxOption'
 
 interface CheckboxState {
   JourneyViewEvent: boolean
